+++ spring-boot-data-jpa-actual/src/main/java/com/fayzal/springboot/datajpaactual/app/models/dao/IClienteDaoPageableAndSorting.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.fayzal.springboot.datajpaactual.app.models.entity.Cliente;
 
-public interface IClienteDaoPageableAndSorting extends PagingAndSortingRepository<Cliente, Integer>{
+public interface IClienteDaoPageableAndSorting extends PagingAndSortingRepository<Cliente, Long>{
